+++ CHANGELOG.md
@@ -17,11 +17,12 @@ since `1.7.0`.
 - Added the Timer class.  A way of handing frame limiting and delta time without depending on libtcod's internals.
 - Added `TCOD_sys_get_internal_context`.  An easy way to get a `TCOD_Context` from the old API.
 - Added `TCOD_sys_get_internal_console`.  This returns a pointer to the root console used by the old API.
-- `TCODColor` can now has implicit conversions to `TCOD_ColorRGB` or `TCOD_ColorRGBA`.
 - Added `tcod::ColorRGB` and `tcod::ColorRGBA` to help with color type conversions.
 - Added libtcod logging functions, for debugging and internal use.
 - Added `tcod::draw_quartergraphics` to replace older C++ blit2x functions.
 - Added `tcod::blit` to replace older C++ blit functions.
+- Added `tcod::stringf` to encapsulate printf formatted strings so that other functions won't have to deal with them.
+  A library like [fmt](https://fmt.dev/latest/index.html) is safer and faster if you have access to it.
 
 ### Changed
 - Initializing libtcod with the old API will now enable VSync by default.